# docs/pappers_immobilier_openapi.yaml
openapi: 3.0.0
info:
    description: |
        L'API de Pappers Immobilier vous permet de récupérer des informations sur les parcelles cadastrales et leurs bases associées (propriétaires, ventes, bâtiments occupants, etc...).

        Vous devez indiquer votre clé d'API dans les requêtes, soit via le header `api-key`, soit (déconseillé) en utilisant le paramètre query `api_token`.

        L'URL d'accès à l'API est https://api-immobilier.pappers.fr/v1/
    version: 1.0.0
    title: Pappers Immobilier API
    contact:
        name: 'Support technique Pappers'
        email: support@pappers.fr
servers:
    - url: https://api-immobilier.pappers.fr/v1/
security:
    - apiKey: []
paths:
    /parcelles/{numero_parcelle}:
        get:
            tags:
                - Parcelles
            summary: 'Récupère une parcelle cadastrale.'
            description: |
                Cette route permet de récupérer les informations disponibles sur la parcelle cadastrale `numero_parcelle`.

                Le paramètre `bases` permet de sélectionner les informations souhaitées dans le retour (propriétaires, ventes, bâtiments occupants, etc...).
            operationId: parcelle
            parameters:
                - name: numero_parcelle
                  in: path
                  required: true
                  description: 'Numéro de parcelle cadastrale.'
                  schema:
                      type: string
                      example: 75101000AG0006
                - $ref: '#/components/parameters/par_page'
                - $ref: '#/components/parameters/page'
                - $ref: '#/components/parameters/curseur'
                - $ref: '#/components/parameters/bases'
                - $ref: '#/components/parameters/champs_supplementaires'
                - $ref: '#/components/parameters/adresse'
                - $ref: '#/components/parameters/code_postal'
                - $ref: '#/components/parameters/code_commune'
                - $ref: '#/components/parameters/region'
                - $ref: '#/components/parameters/departement'
                - $ref: '#/components/parameters/latitude'
                - $ref: '#/components/parameters/longitude'
                - $ref: '#/components/parameters/distance'
                - $ref: '#/components/parameters/type_syndic_copropriete'
                - $ref: '#/components/parameters/siret_syndic_professionnel_copropriete'
                - $ref: '#/components/parameters/nom_copropriete'
                - $ref: '#/components/parameters/nombre_lots_copropriete_min'
                - $ref: '#/components/parameters/nombre_lots_copropriete_max'
                - $ref: '#/components/parameters/nombre_coproprietes_min'
                - $ref: '#/components/parameters/nombre_coproprietes_max'
                - $ref: '#/components/parameters/siren_proprietaire'
                - $ref: '#/components/parameters/tranche_effectif_proprietaire_min'
                - $ref: '#/components/parameters/tranche_effectif_proprietaire_max'
                - $ref: '#/components/parameters/categorie_juridique_proprietaire'
                - $ref: '#/components/parameters/code_naf_proprietaire'
                - $ref: '#/components/parameters/denomination_proprietaire'
                - $ref: '#/components/parameters/nombre_proprietaires_min'
                - $ref: '#/components/parameters/nombre_proprietaires_max'
                - $ref: '#/components/parameters/siren_occupant'
                - $ref: '#/components/parameters/tranche_effectif_occupant_min'
                - $ref: '#/components/parameters/tranche_effectif_occupant_max'
                - $ref: '#/components/parameters/categorie_juridique_occupant'
                - $ref: '#/components/parameters/code_naf_occupant'
                - $ref: '#/components/parameters/denomination_occupant'
                - $ref: '#/components/parameters/nature_vente'
                - $ref: '#/components/parameters/nombre_occupants_min'
                - $ref: '#/components/parameters/nombre_occupants_max'
                - $ref: '#/components/parameters/nature_vente'
                - $ref: '#/components/parameters/type_local_vente'
                - $ref: '#/components/parameters/date_vente_min'
                - $ref: '#/components/parameters/date_vente_max'
                - $ref: '#/components/parameters/prix_vente_min'
                - $ref: '#/components/parameters/prix_vente_max'
                - $ref: '#/components/parameters/surface_bati_vente_min'
                - $ref: '#/components/parameters/surface_bati_vente_max'
                - $ref: '#/components/parameters/surface_terrain_vente_min'
                - $ref: '#/components/parameters/surface_terrain_vente_max'
                - $ref: '#/components/parameters/nombre_pieces_vente_min'
                - $ref: '#/components/parameters/nombre_pieces_vente_max'
                - $ref: '#/components/parameters/nombre_ventes_min'
                - $ref: '#/components/parameters/nombre_ventes_max'
                - $ref: '#/components/parameters/annee_construction_batiment_min'
                - $ref: '#/components/parameters/annee_construction_batiment_max'
                - $ref: '#/components/parameters/nombre_logements_batiment_min'
                - $ref: '#/components/parameters/nombre_logements_batiment_max'
                - $ref: '#/components/parameters/surface_batiment_min'
                - $ref: '#/components/parameters/surface_batiment_max'
                - $ref: '#/components/parameters/nature_batiment'
                - $ref: '#/components/parameters/usage_batiment'
                - $ref: '#/components/parameters/nombre_batiments_min'
                - $ref: '#/components/parameters/nombre_batiments_max'
                - $ref: '#/components/parameters/classe_bilan_dpe'
                - $ref: '#/components/parameters/type_installation_chauffage_dpe'
                - $ref: '#/components/parameters/type_energie_chauffage_dpe'
                - $ref: '#/components/parameters/prix_fonds_de_commerce_min'
                - $ref: '#/components/parameters/prix_fonds_de_commerce_max'
                - $ref: '#/components/parameters/date_fonds_de_commerce_min'
                - $ref: '#/components/parameters/date_fonds_de_commerce_max'
                - $ref: '#/components/parameters/code_naf_fonds_de_commerce'
                - $ref: '#/components/parameters/nombre_fonds_de_commerce_min'
                - $ref: '#/components/parameters/nombre_fonds_de_commerce_max'
                - $ref: '#/components/parameters/statut_permis'
                - $ref: '#/components/parameters/date_autorisation_permis_min'
                - $ref: '#/components/parameters/date_autorisation_permis_max'
                - $ref: '#/components/parameters/nombre_permis_min'
                - $ref: '#/components/parameters/nombre_permis_max'

            responses:
                '200':
                    description: Parcelle trouvée. Le retour contient un objet décrivant la parcelle.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ParcelleFiche'
                '400':
                    description: Paramètres de la requête incorrects.
                '401':
                    description: Clé API incorrecte.
                '404':
                    description: Parcelle inexistante.
...
components:
    parameters:
        bases:
            name: bases
            in: query
            description: "Bases de données souhaitées dans le retour. Il est possible d’indiquer plusieurs bases en les séparant par des virgules. Si aucune base n’est spécifiée, seules les informations sur la parcelle seront retournées (1 crédit). Coût par base : `ventes` 2 crédits, `proprietaires` 2 crédits, `occupants` 2 crédits, `permis` 1 crédit, `fonds_de_commerce` 1 crédit, `batiments` 1 crédit, `dpe` 1 crédit, `coproprietes` 1 crédit, `amenagements` 1 crédit, `documents_urbanisme` 1 crédit."
            required: false
            schema:
                type: string
                enum: [proprietaires, ventes, batiments, dpe, occupants, permis, fonds_de_commerce, coproprietes, amenagements, documents_urbanisme]
                example: ventes, proprietaires
        par_page:
            name: par_page
            in: query
            description: 'Nombre de résultats affichés sur une page. Valeur par défaut : `10`.'
            required: false
            schema:
                type: integer
                example: 20
        page:
            name: page
            in: query
            description: 'Page de résultats. Valeur par défaut : `1`.'
            required: false
            schema:
                type: integer
                example: 1
        curseur:
            name: curseur
            in: query
            description: "Curseur servant à parcourir l'ensemble des résultats (alternativement à la pagination qui est limitée à 400 résultats maximum). Doit valoir `*` pour la première requête, et doit pour les requêtes suivantes reprendre la valeur `curseurSuivant` retournée par la dernière réponse."
            required: false
            schema:
                type: string
                example: '*'
        champs_supplementaires:
            name: champs_supplementaires
            in: query
            description: |
                Liste des champs supplémentaires à inclure dans le retour. Certains champs peuvent entraîner une consommation de jetons supplémentaires.

                Champs supplémentaires disponibles :
                - `tous` : 3 jetons supplémentaires
                - `bounding_box` : 1 jeton supplémentaire
                - `arpente` : gratuit
                - `adresse` : 1 jeton supplémentaire
                - `ventes.code_type_local` : gratuit
                - `ventes.nature_culture` : gratuit
                - `ventes.ancienne_parcelle_cadastrale` : gratuit
                - `proprietaires.parcelles` : 1 jeton supplémentaire
                - `proprietaires.personnes_physiques` : 1 jeton supplémentaire
                - `proprietaires.representants_personnes_morales` : 1 jeton supplémentaire
                - `occupants.categorie_juridique` : gratuit
                - `occupants.personnes_physiques` : 1 jeton supplémentaire
                - `occupants.representants_personnes_morales` : 1 jeton supplémentaire
                - `permis.zone_operatoire` : gratuit
                - `permis.complet` : 1 jeton supplémentaire

                Champs de contact :
                - `occupants.emails` : 3 crédits supplémentaires
                - `occupants.telephones` : 3 crédits supplémentaires
                - `occupants.sites_internet` : 1 crédit supplémentaire
                - `occupants.lien_x` : 3 crédits supplémentaires
                - `occupants.lien_linkedin` : 3 crédits supplémentaires
                - `occupants.lien_facebook` : 3 crédits supplémentaires
                - `coproprietes.syndic_professionnel.emails` : 3 crédits supplémentaires
                - `coproprietes.syndic_professionnel.telephones` : 3 crédits supplémentaires
                - `coproprietes.syndic_professionnel.sites_internet` : 1 crédit supplémentaire
                - `coproprietes.syndic_professionnel.lien_x` : 3 crédits supplémentaires
                - `coproprietes.syndic_professionnel.lien_linkedin` : 3 crédits supplémentaires
                - `coproprietes.syndic_professionnel.lien_facebook` : 3 crédits supplémentaires
                - `coproprietes.representant_legal.emails` : 3 crédits supplémentaires
                - `coproprietes.representant_legal.telephones` : 3 crédits supplémentaires
                - `coproprietes.representant_legal.sites_internet` : 1 crédit supplémentaire
                - `coproprietes.representant_legal.lien_x` : 3 crédits supplémentaires
                - `coproprietes.representant_legal.lien_linkedin` : 3 crédits supplémentaires
                - `coproprietes.representant_legal.lien_facebook` : 3 crédits supplémentaires
                - `proprietaires.emails` : 3 crédits supplémentaires
                - `proprietaires.telephones` : 3 crédits supplémentaires
                - `proprietaires.sites_internet` : 1 crédit supplémentaire
                - `proprietaires.lien_x` : 3 crédits supplémentaires
                - `proprietaires.lien_linkedin` : 3 crédits supplémentaires
                - `proprietaires.lien_facebook` : 3 crédits supplémentaires
                - `fonds_de_commerce.acheteur.emails` : 3 crédits supplémentaires
                - `fonds_de_commerce.acheteur.telephones` : 3 crédits supplémentaires
                - `fonds_de_commerce.acheteur.sites_internet` : 1 crédit supplémentaire
                - `fonds_de_commerce.acheteur.lien_x` : 3 crédits supplémentaires
                - `fonds_de_commerce.acheteur.lien_linkedin` : 3 crédits supplémentaires
                - `fonds_de_commerce.acheteur.lien_facebook` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_proprietaire.emails` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_proprietaire.telephones` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_proprietaire.sites_internet` : 1 crédit supplémentaire
                - `fonds_de_commerce.precedent_proprietaire.lien_x` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_proprietaire.lien_linkedin` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_proprietaire.lien_facebook` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_exploitant.emails` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_exploitant.telephones` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_exploitant.sites_internet` : 1 crédit supplémentaire
                - `fonds_de_commerce.precedent_exploitant.lien_x` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_exploitant.lien_linkedin` : 3 crédits supplémentaires
                - `fonds_de_commerce.precedent_exploitant.lien_facebook` : 3 crédits supplémentaires
                - `personnes_physiques.emails` : 7 crédits supplémentaires
                - `emails` : 3 crédits supplémentaires (toutes les bases)
                - `telephones` : 3 crédits supplémentaires (toutes les bases)
                - `sites_internet` : 1 crédit supplémentaire (toutes les bases)
                - `lien_x` : 3 crédits supplémentaires (toutes les bases)
                - `lien_linkedin` : 3 crédits supplémentaires (toutes les bases)
                - `lien_facebook` : 3 crédits supplémentaires (toutes les bases)

            required: false
            schema:
                type: string
                example: tous
...
        adresse:
            name: adresse
            in: query
            description: 'Adresse.'
            required: false
            schema:
                type: string
                example: 10 Rue Ordener 75018 Paris
        code_postal:
            name: code_postal
            in: query
            description: 'Code postal.'
            required: false
            schema:
                type: string
                example: 75001
        code_commune:
            name: code_commune
            in: query
            description: 'Code commune.'
            required: false
            schema:
                type: string
                example: 75101
        region:
            name: region
            in: query
            description: Code de la région de l'un des établissements de la parcelle, selon la [nomenclature Insee](https://www.insee.fr/fr/information/4316069#titre-bloc-18). Il est possible d'indiquer plusieurs codes régions en les séparant par des virgules.
            required: false
            schema:
                type: string
                example: 11
        departement:
            name: departement
            in: query
            description: "Numéro de département. Il est possible d'indiquer plusieurs départements en les séparant par des virgules."
            required: false
            schema:
                type: string
                example: 93
        latitude:
            name: latitude
            in: query
            description: 'Latitude. La latitude, la longitude et la distance doivent être renseignées ensemble.'
            required: false
            schema:
                type: number
                example: 48.86
        longitude:
            name: longitude
            in: query
            description: 'Longitude. La latitude, la longitude et la distance doivent être renseignées ensemble.'
            required: false
            schema:
                type: number
                example: 2.35
        distance:
            name: distance
            in: query
            description: 'Distance en mètres. La latitude, la longitude et la distance doivent être renseignées ensemble.'
            required: false
            schema:
                type: integer
                example: 20
        type_syndic_copropriete:
            name: type_syndic_copropriete
            in: query
            description: 'Type de syndic de copropriété.'
            required: false
            schema:
                type: string
                enum: [professionnel, benevole]
                example: professionnel
        siret_syndic_professionnel_copropriete:
            name: siret_syndic_professionnel_copropriete
            in: query
            description: 'SIRET du syndic professionnel de copropriété.'
            required: false
            schema:
                type: string
                example: 96550338600074
        nom_copropriete:
            name: nom_copropriete
            in: query
            description: 'Nom de la copropriété.'
            required: false
            schema:
                type: string
                example: enedis
        nombre_lots_copropriete_min:
            name: nombre_lots_copropriete_min
            in: query
            description: 'Nombre de lots minimum de la copropriété.'
            required: false
            schema:
                type: integer
                example: 1
        nombre_lots_copropriete_max:
            name: nombre_lots_copropriete_max
            in: query
            description: 'Nombre de lots maximum de la copropriété.'
            required: false
            schema:
                type: integer
                example: 2
        nombre_coproprietes_min:
            name: nombre_coproprietes_min
            in: query
            description: 'Nombre de copropriété minimum.'
            required: false
            schema:
                type: integer
                example: 0
        nombre_coproprietes_max:
            name: nombre_coproprietes_max
            in: query
            description: 'Nombre de copropriété maximum.'
            required: false
            schema:
                type: integer
                example: 1
        siren_proprietaire:
            name: siren_proprietaire
            in: query
            description: 'SIREN du propriétaire.'
            required: false
            schema:
                type: string
                example: 444608442
        tranche_effectif_proprietaire_min:
            name: tranche_effectif_proprietaire_min
            in: query
            description: "Tranche d'effectifs minimale de l'entreprise, selon la [nomenclature Sirene](https://www.sirene.fr/sirene/public/variable/tefen).\n\n**Note** : 00 ou NN donneront les mêmes résultats et veulent dire non employeur"
            required: false
            schema:
                type: string
                example: 01
        tranche_effectif_proprietaire_max:
            name: tranche_effectif_proprietaire_max
            in: query
            description: "Tranche d'effectifs maximale de l'entreprise, selon la [nomenclature Sirene](https://www.sirene.fr/sirene/public/variable/tefen).\n\n**Note** : 00 ou NN donneront les mêmes résultats et veulent dire non employeur"
            required: false
            schema:
                type: string
                example: 02
        categorie_juridique_proprietaire:
            name: categorie_juridique_proprietaire
            in: query
            description: "Catégorie juridique de l'entreprise, selon la [nomenclature Insee](https://www.insee.fr/fr/information/2028129). \n\n**Note** : Le code correspond à celui de l'INSEE, à l'exception des SASU qui auront comme code 5720 et les EURL qui auront comme code 5498."
            required: false
            schema:
                type: string
                example: 6540
        code_naf_proprietaire:
            name: code_naf_proprietaire
            in: query
            description: Code NAF de l'entreprise. Il est possible d'indiquer plusieurs codes NAF en les séparant par des virgules.
            required: false
            schema:
                type: string
                example: 52.10B
        denomination_proprietaire:
            name: denomination_proprietaire
            in: query
            description: "Dénomination de l'entreprise."
            required: false
            schema:
                type: string
                example: enedis
        nombre_proprietaires_min:
            name: nombre_proprietaires_min
            in: query
            description: 'Nombre de proprietaires minimum.'
            required: false
            schema:
                type: integer
                example: 0
        nombre_proprietaires_max:
            name: nombre_proprietaires_max
            in: query
            description: 'Nombre de proprietaires maximum.'
            required: false
            schema:
                type: integer
                example: 1
        siren_occupant:
            name: siren_occupant
            in: query
            description: "SIREN de l'occupant."
            required: false
            schema:
                type: string
                example: 444608442
        tranche_effectif_occupant_min:
            name: tranche_effectif_occupant_min
            in: query
            description: "Tranche d'effectifs minimale de l'entreprise, selon la [nomenclature Sirene](https://www.sirene.fr/sirene/public/variable/tefen).\n\n**Note** : 00 ou NN donneront les mêmes résultats et veulent dire non employeur"
            required: false
            schema:
                type: string
                example: 01
        tranche_effectif_occupant_max:
            name: tranche_effectif_occupant_max
            in: query
            description: "Tranche d'effectifs maximale de l'entreprise, selon la [nomenclature Sirene](https://www.sirene.fr/sirene/public/variable/tefen).\n\n**Note** : 00 ou NN donneront les mêmes résultats et veulent dire non employeur"
            required: false
            schema:
                type: string
                example: 02
        categorie_juridique_occupant:
            name: categorie_juridique_occupant
            in: query
            description: "Catégorie juridique de l'entreprise, selon la [nomenclature Insee](https://www.insee.fr/fr/information/2028129). \n\n**Note** : Le code correspond à celui de l'INSEE, à l'exception des SASU qui auront comme code 5720 et les EURL qui auront comme code 5498."
            required: false
            schema:
                type: string
                example: 6540
        code_naf_occupant:
            name: code_naf_occupant
            in: query
            description: Code NAF de l'entreprise. Il est possible d'indiquer plusieurs codes NAF en les séparant par des virgules.
            required: false
            schema:
                type: string
                example: 52.10B
        denomination_occupant:
            name: denomination_occupant
            in: query
            description: "Dénomination de l'entreprise."
            required: false
            schema:
                type: string
                example: enedis
        nombre_occupants_min:
            name: nombre_occupants_min
            in: query
            description: "Nombre d'occupants minimum de la parcelle."
            required: false
            schema:
                type: integer
                example: 0
        nombre_occupants_max:
            name: nombre_occupants_max
            in: query
            description: "Nombre d'occupants maximum de la parcelle."
            required: false
            schema:
                type: integer
                example: 2
        nature_vente:
            name: nature_vente
            in: query
            description: 'Nature de la mutation.'
            required: false
            schema:
                type: string
                enum: [vente, echange, adjudication, expropriation, vente_futur_achevement, vente_terrain_batir]
                example: vente
        type_local_vente:
            name: type_local_vente
            in: query
            description: 'Type de local.'
            required: false
            schema:
                type: string
                enum: [appartement, maison, dependance, local_industriel_commercial_ou_assimile]
                example: maison
        date_vente_min:
            name: date_vente_min
            in: query
            description: 'Date de mutation minimale, au format AAAA-MM-JJ.'
            required: false
            schema:
                type: string
                example: 2020-01-01
        date_vente_max:
            name: date_vente_max
            in: query
            description: 'Date de mutation maximale, au format AAAA-MM-JJ.'
            required: false
            schema:
                type: string
                example: 2024-01-01
        prix_vente_min:
            name: prix_vente_min
            in: query
            description: 'Prix de vente minimum.'
            required: false
            schema:
                type: integer
                example: 0
        prix_vente_max:
            name: prix_vente_max
            in: query
            description: 'Prix de vente maximum.'
            required: false
            schema:
                type: integer
                example: 100000
        surface_bati_vente_min:
            name: surface_bati_vente_min
            in: query
            description: 'Surface du bâti minimale.'
            required: false
            schema:
                type: integer
                example: 50
        surface_bati_vente_max:
            name: surface_bati_vente_max
            in: query
            description: 'Surface du bâti maximale.'
            required: false
            schema:
                type: integer
                example: 100
        surface_terrain_vente_min:
            name: surface_terrain_vente_min
            in: query
            description: 'Surface du terrain minimale.'
            required: false
            schema:
                type: integer
                example: 0
        surface_terrain_vente_max:
            name: surface_terrain_vente_max
            in: query
            description: 'Surface du terrain maximale.'
            required: false
            schema:
                type: integer
                example: 100
        nombre_pieces_vente_min:
            name: nombre_pieces_vente_min
            in: query
            description: 'Nombre de pièces minimum.'
            required: false
            schema:
                type: integer
                example: 0
        nombre_pieces_vente_max:
            name: nombre_pieces_vente_max
            in: query
            description: 'Nombre de pièces maximum.'
            required: false
            schema:
                type: integer
                example: 1
        nombre_ventes_min:
            name: nombre_ventes_min
            in: query
            description: 'Nombre de ventes minimum de la parcelle.'
            required: false
            schema:
                type: integer
                example: 0
        nombre_ventes_max:
            name: nombre_ventes_max
            in: query
            description: 'Nombre de ventes maximum de la parcelle.'
            required: false
            schema:
                type: integer
                example: 2
        annee_construction_batiment_min:
            name: annee_construction_batiment_min
            in: query
            description: 'Année de construction minimale du bâtiment.'
            required: false
            schema:
                type: integer
                example: 1900
        annee_construction_batiment_max:
            name: annee_construction_batiment_max
            in: query
            description: 'Année de construction maximale du bâtiment.'
            required: false
            schema:
                type: integer
                example: 1990
        nombre_logements_batiment_min:
            name: nombre_logements_batiment_min
            in: query
            description: 'Nombre de logements minimum du bâtiment.'
            required: false
            schema:
                type: integer
                example: 0
        nombre_logements_batiment_max:
            name: nombre_logements_batiment_max
            in: query
            description: 'Nombre de logements maximum du bâtiment.'
            required: false
            schema:
                type: integer
                example: 100
        surface_batiment_min:
            name: surface_batiment_min
            in: query
            description: 'Surface du bâtiment minimale.'
            required: false
            schema:
                type: integer
                example: 0
        surface_batiment_max:
            name: surface_batiment_max
            in: query
            description: 'Surface du bâtiment maximale.'
            required: false
            schema:
                type: integer
                example: 1000
        nature_batiment:
            name: nature_batiment
            in: query
            description: 'Nature du bâtiment.'
            required: false
            schema:
                type: string
                enum:
                    [
                        indifferenciee,
                        silo,
                        industriel_agricole_ou_commercial,
                        monument,
                        chapelle,
                        chateau,
                        tour_donjon,
                        fort_blockhaus_casemate,
                        eglise,
                        moulin_a_vent,
                        arene_ou_theatre_antique,
                        arc_de_triomphe,
                        tribune
                    ]
                example: indifferenciee
        usage_batiment:
            name: usage_batiment
            in: query
            description: 'Usage du bâtiment.'
            required: false
            schema:
                type: string
                enum:
                    [
                        tertiaire_et_autres,
                        indifferencie,
                        annexe,
                        residentiel,
                        residentiel_individuel,
                        residentiel_collectif,
                        commercial_et_services,
                        agricole,
                        industriel,
                        sportif,
                        religieux,
                        dependance,
                        secondaire
                    ]
                example: residentiel
        nombre_batiments_min:
            name: nombre_batiments_min
            in: query
            description: 'Nombre de batiments minimum.'
            required: false
            schema:
                type: integer
                example: 0
        nombre_batiments_max:
            name: nombre_batiments_max
            in: query
            description: 'Nombre de batiments maximum.'
            required: false
            schema:
                type: integer
                example: 1
        classe_bilan_dpe:
            name: classe_bilan_dpe
            in: query
            description: 'Classe du bilan DPE.'
            required: false
            schema:
                type: string
                enum: [A, B, C, D, E, F, G]
                example: A
        type_installation_chauffage_dpe:
            name: type_installation_chauffage_dpe
            in: query
            description: "Type d'installation de chauffage."
            required: false
            schema:
                type: string
                enum: [individuel, collectif]
                example: collectif
        type_energie_chauffage_dpe:
            name: type_energie_chauffage_dpe
            in: query
            description: "Type d'énergie de chauffage."
            required: false
            schema:
                type: string
                enum: [gaz, electricite, fioul, bois, gpl_butane_propane, solaire, charbon, reseau_de_chaleur]
                example: gaz
        prix_fonds_de_commerce_min:
            name: prix_fonds_de_commerce_min
            in: query
            description: 'Prix de fonds de commerce minimum.'
            required: false
            schema:
                type: integer
                example: 150000
        prix_fonds_de_commerce_max:
            name: prix_fonds_de_commerce_max
            in: query
            description: 'Prix de fonds de commerce maximum.'
            required: false
            schema:
                type: integer
                example: 200000
...
    securitySchemes:
        apiKey:
            description: Clé API fournie dans votre espace membre.
            type: apiKey
            name: api-key
            in: header